// new-frontend/foresty/src/app/models/response/response.tree-property.ts
export interface TreeProperty {
    id:                 number;
    taxationYear:       number;
    age:                number;
    jar:                null;
    pokolenie:          null;
    formaKrHorizontal:  null;
    formaKrVertical:    null;
    height:             number;
    heightNaibDiametra: number | null;
    crownLength:        number | null;
    diametrNs:          number;
    diametrWe:          number;
    crownDiametrNs:     number | null;
    crownDiametrWe:     number | null;
    idSuitability:      number;
    idCraft:            number;
    idTree:             number;
    idTaxationYears:    number;
    craftName:          null;
    year:               number;
    suitabilityName:    string;
}